Fix DeviceInfo.getUniqueID() returning wrong value on iOS sometimes
Sometimes (I'm not sure in what circumstances) DeviceInfo.getUniqueID() returns "00000000-0000-0000-0000-000000000000" on iOS. This means it is not unique because it returns the same value on different devices. This commit checks for that bad value and continues in the chain of fallbacks when uid has that bad value.
Possibly related links:
https://gist.github.com/miguelcma/e8f291e54b025815ca46#gistcomment-1815156
https://stackoverflow.com/questions/12605257/the-advertisingidentifier-and-identifierforvendor-return-00000000-0000-0000-000